CVhomework_2/requirements.txt
matplotlib==3.3.2
scipy==1.5.2
numpy==1.19.2
opencv_python==4.4.0.46
Pillow==8.0.1
